requirements.txt
gradio>=3.50.0
transformers>=4.37.0
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
gtts>=2.3.2
Pillow>=10.0.0
sentencepiece>=0.1.99
protobuf<4.0.0
accelerate>=0.21.0

